# Use each with object instead of reduce to improve readability and performance