Slightly simplifies the code and gives a minor performance improvement
on newer Ruby versions

M1 Pro, Ruby 3.3.5 x64, single char:

```
 name_to_indices_new:  1370019.0 i/s
     name_to_indices:  1322752.7 i/s - 1.04x  (± 0.00) slower
```

M1 Pro, Ruby 3.3.3 ARM, single char:

```
 name_to_indices_new:  2324745.0 i/s
     name_to_indices:  2157339.6 i/s - 1.08x  (± 0.00) slower
```